refactor(firmware): decompose FirmwareService into services/firmware/
The module sat at exactly 1000 of its 1000 code lines, so the next change
to the BIOS surface failed the size gate before it started.

Split along the seams the class already had, with the façade keeping the
public surface unchanged:

- listing.py   the RomM listing and its TTL cache, in memory and in SQLite
- demand.py    the resolver reading, each file's destination, and presence
- status.py    the System page and the per-game BIOS check
- downloads.py the three entry points and the downloaded_bios record
- deletion.py  Delete BIOS, driven by those records alone

demand.py is the shared kernel rather than a fifth seam: the status
surfaces and the download paths both need the destination and the
presence answer, and letting the read path reach into the downloader for
them would make those builders part of the downloader's public surface.

Behaviour is unchanged; the methods a peer now calls lose their leading
underscore, as the services rule requires.
